refactor: remove cycle from Profile
